+++ src/components/card/ComponentDefinitionCard.vue
@@ -52,14 +52,12 @@
 <script setup>
 import { computed, ref, toRef } from 'vue';
 import {
-  addNewComponent,
   addNewTemplateComponent,
   getPluginByName,
   renderModel,
 } from 'src/composables/PluginManager';
 import { useRoute } from 'vue-router';
 import DefinitionMenu from 'components/menu/DefinitionMenu.vue';
-import { getModelFiles } from 'src/composables/Project';
 
 const route = useRoute();
 const props = defineProps({
@@ -117,17 +115,10 @@ const componentIcon = computed(() => {
  */
 async function onClickItem() {
   if (!props.definition.isTemplate) {
-    let path = (await getModelFiles(projectName.value, query.value.path, plugin.value))[0]?.path;
-
-    if (!path) {
-      path = `${query.value.path}/${plugin.value.configuration.defaultFileName}`;
-    }
-
-    await addNewComponent(
-      projectName.value,
-      plugin.value,
-      path,
+    plugin.value.addComponent(
+      'root',
       props.definition,
+      query.value.path,
     );
   } else {
     await addNewTemplateComponent(