+++ changelog.md
@@ -19,12 +19,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Improve unit tests coverage.
 * Technical content:
   * Update import of vue components.
+* Plugin management:
+  * Work with [plugin-core](https://github.com/ditrit/leto-modelizer-plugin-core/tree/0.22.0) version 0.22.0.
+  * Work with [terrator-plugin](https://github.com/ditrit/terrator-plugin/tree/0.7.1) version 0.7.1.
+  * Work with [githubator-plugin](https://github.com/ditrit/githubator-plugin/tree/0.3.0) version 0.3.0.
 
 ### Fixed
 
 * Fix bug on click on component definition from plugins with single file diagrams, see [this issue](https://github.com/ditrit/leto-modelizer/issues/407).
 * Fix bug about rewritten file on checkout, see [this issue](https://github.com/ditrit/leto-modelizer/issues/422).
 * Fix bug about required management for RegExp rule, see [this issue](https://github.com/ditrit/leto-modelizer/issues/387).
+* Fix bugs about component creation on a click or a drag&drop, see [this issue](https://github.com/ditrit/leto-modelizer/issues/401) and [this issue](https://github.com/ditrit/leto-modelizer/issues/415).
 
 ## [1.3.0] - 2023/09/12